+++ init.lua
@@ -1,5 +1,6 @@
 require('impatient') -- This needs to be first
 require('packer_commands')
+require('globals')
 require('config')
 require('settings')
 require('colorscheme')
@@ -7,4 +8,5 @@ require('keymappings')
 require('autocmds')
 require('functions')
 require('lsp.config')
-require('globals')
+require('lsp.run')
+require('lsp.functions')