+++ lua/colorscheme.lua
@@ -31,6 +31,7 @@ if EcoVim.colorscheme == 'tokyonight' then
   vim.highlight.create('StatusLine', { guibg = "None" }, false);
   vim.highlight.create('StatusLineNC', { guibg = "None" }, false);
   vim.highlight.create('rainbowcol1', { guifg = tokyonight_colors.blue, ctermfg = 9 }, false);
+  vim.highlight.create('Boolean', { guifg = "#F7768E" }, false);
 
   -- Completion Menu Colors
 
@@ -76,5 +77,5 @@ if EcoVim.colorscheme == 'nightfly' then
   vim.highlight.create('Comment', { gui = "italic", cterm = "italic" }, false);
 
   -- Change color of floating popup border
-  vim.highlight.link('FloatBorder', EcoVim.ui.float.highlight, true);
+  vim.highlight.link('FloatBorder', 'NightflyRed', true);
 end