+++ lua/config.lua
@@ -12,8 +12,7 @@ EcoVim = {
   colorscheme = 'tokyonight', -- nightfly/tokyonight
   ui = {
     float = {
-      border = 'rounded',
-      highlight = 'NightflyRed' -- check available by :Telescope highlights
+      border = 'rounded'
     }
   },
   plugins = {
@@ -40,23 +39,6 @@ EcoVim = {
   statusline = {
     path_enabled = true,
     path = 'relative' -- absolute/relative
-  },
-  -- LSP settings (for overriding per client)
-  -- Load servers (They will be automatically installed after next "Sync plugins" launch)
-  -- Check installed servers by :LspInstallInfo
-  lsp = {
-    servers = {
-      'bash',
-      'css',
-      'eslint',
-      'graphql',
-      'html',
-      'json',
-      'lua',
-      -- 'tailwind',
-      'tsserver',
-      -- 'vue2'
-    }
   }
 }