+++ lua/plugins/tree.lua
@@ -1,4 +1,20 @@
-vim.g.nvim_tree_respect_buf_cwd = 1
+vim.g.nvim_tree_respect_buf_cwd = 1 --0 by default, will change cwd of nvim-tree to that of new buffer's when opening nvim-tree
+vim.g.nvim_tree_group_empty     = 1 --0 by default, compact folders that only contain a single folder into one node in the file tree
+vim.g.nvim_tree_git_hl          = 1 --0 by default, will enable file highlight for git attributes
+
+local git_icons = {
+  unstaged = "",
+  staged = "",
+  unmerged = "",
+  renamed = "➜",
+  untracked = "",
+  deleted = "",
+  ignored = "◌"
+}
+
+vim.g.nvim_tree_icons = {
+  git = git_icons
+}
 
 local keymappings = {
   { key = {"<CR>", "o", "l", "<2-LeftMouse>"}, action = "edit" },
@@ -49,8 +65,6 @@ require'nvim-tree'.setup {
   open_on_setup       = false,
   -- will not open on setup if the filetype is in this list
   ignore_ft_on_setup  = {},
-  -- closes neovim automatically when the tree is the last **WINDOW** in the view
-  auto_close          = false,
   -- opens the tree when changing/opening a new tab if the tree wasn't previously opened
   open_on_tab         = false,
   -- hijack the cursor in the tree to put it at the start of the filename
@@ -83,7 +97,7 @@ require'nvim-tree'.setup {
   -- configuration options for the system open command (`s` in the tree by default)
   system_open = {
     -- the command to run this, leaving nil should work in most cases
-    cmd  = nil,
+    cmd  = "",
     -- the command arguments as a list
     args = {}
   },
@@ -96,14 +110,33 @@ require'nvim-tree'.setup {
     ignore = true,
     timeout = 500,
   },
+  actions = {
+    use_system_clipboard = true,
+    change_dir = {
+      enable = true,
+      global = false,
+      restrict_above_cwd = false,
+    },
+    open_file = {
+      quit_on_open = false,
+      -- if true the tree will resize itself after opening a file
+      resize_window = true,
+      window_picker = {
+        enable = true,
+        chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890",
+        exclude = {
+          filetype = { "notify", "packer", "qf", "diff", "fugitive", "fugitiveblame" },
+          buftype = { "nofile", "terminal", "help" },
+        },
+      },
+    },
+  },
   view = {
     -- width of the window, can be either a number (columns) or a string in `%`
     width = 40,
     hide_root_folder = false,
     -- side of the tree, can be one of 'left' | 'right' | 'top' | 'bottom'
     side = 'left',
-    -- if true the tree will resize itself after opening a file
-    auto_resize = false,
     mappings = {
       -- custom only false will merge the list with the default mappings
       -- if true, it will only use your list to set the mappings
@@ -111,8 +144,8 @@ require'nvim-tree'.setup {
       -- list of mappings to set on the tree manually
       list = keymappings
     },
-    number = true,
-    relativenumber = true
+    number = false,
+    relativenumber = false
   },
   trash = {
     cmd = "trash",
@@ -127,5 +160,3 @@ require'nvim-tree'.setup {
 
 vim.api.nvim_set_keymap("n", "<C-e>", "<cmd>lua require'nvim-tree'.toggle()<CR>", {noremap = true, silent = true})
 vim.api.nvim_set_keymap("n", ",e", "<cmd>lua require'nvim-tree'.toggle()<CR>", {noremap = true, silent = true})
--- vim.api.nvim_set_keymap("n", ",\\", "<cmd>lua require'nvim-tree'.toggle()<CR>", {noremap = true, silent = true})
-